link download

+++ storage-application/src/main/java/digital/container/service/taxdocument/DownloadTaxDocumentService.java
@@ -1,9 +1,11 @@
 package digital.container.service.taxdocument;
 
+import digital.container.service.download.LinkDownloadService;
 import digital.container.storage.domain.model.download.LinkDownload;
 import digital.container.storage.domain.model.file.*;
 import digital.container.util.GenerateHash;
 import digital.container.storage.domain.model.util.SearchScheduling;
+import digital.container.util.LocalFileUtil;
 import digital.container.util.ZipUtil;
 import org.apache.commons.io.FileUtils;
 import org.hibernate.Hibernate;
@@ -30,6 +32,9 @@ public class DownloadTaxDocumentService {
     @Autowired
     private SearchTaxDocumentService searchTaxDocumentService;
 
+    @Autowired
+    private LinkDownloadService linkDownloadService;
+
     public LinkDownload generateLinkToDownload(SearchScheduling searchScheduling) {
         String nameZip = "download";
         String path = this.generateZip(searchScheduling, nameZip);
@@ -37,10 +42,13 @@ public class DownloadTaxDocumentService {
         LinkDownload linkDownload = new LinkDownload();
         linkDownload.setHash(GenerateHash.generateDownload());
         linkDownload.generateDescription(searchScheduling);
-        linkDownload.setName(nameZip);
+        linkDownload.setName(nameZip + ".zip");
         linkDownload.setRelativePath(path);
 
-        return linkDownload;
+        File file = new File(linkDownload.getRelativePath() + linkDownload.getName());
+        linkDownload.setSize(file.length());
+
+        return this.linkDownloadService.save(linkDownload);
     }
 
     private String generateZip(SearchScheduling searchScheduling, String nameZip) {
@@ -52,7 +60,14 @@ public class DownloadTaxDocumentService {
                     if(file instanceof DatabaseFile) {
                         DatabaseFile df = (DatabaseFile) file;
 
-                        String pathFile = path + getFileName(file);
+                        String pathFile = null;
+
+                        if(df.getRelativePath() != null) {
+                            pathFile = path + df.getRelativePath().substring(0, df.getRelativePath().lastIndexOf('/')+1);
+                        } else {
+                            pathFile = path + getFileName(file);
+                        }
+
                         new File(pathFile).mkdirs();
 
                         File outputFile = new File(pathFile + df.getName());
@@ -74,36 +89,44 @@ public class DownloadTaxDocumentService {
 
                     } else {
                         LocalFile lf = (LocalFile) file;
-
+                        try {
+                            FileUtils.copyFile(new File(LocalFileUtil.DIRECTORY_PATH + "/" + lf.getRelativePath()), new File(path + lf.getRelativePath()));
+                        } catch (IOException e) {
+                            LOGGER.error("Erro ao tentar copiar o arquivo para a pasta de download." + e.getMessage());
+                            e.printStackTrace();
+                        }
                     }
                 });
 
+        if(taxDocumentBySearchScheduling.size() > 0 ) {
+            ZipOutputStream zipOutputStream = null;
+            String zipFileName = getDirectoryPath();
+            new File(zipFileName).mkdirs();
 
-        ZipOutputStream zipOutputStream = null;
-        String zipFileName = getDirectoryPath();
-        new File(zipFileName).mkdirs();
-
-        try {
-
-            zipOutputStream = new ZipOutputStream(new FileOutputStream(zipFileName+nameZip+".zip"));
-            ZipUtil.zipFolder(new File(path), zipOutputStream, "");
-        } catch (Exception e) {
-            LOGGER.error("Erro ao tentar zipar os arquivos."+e.getMessage());
-        } finally {
             try {
-                zipOutputStream.close();
-            } catch (IOException e) {
-                LOGGER.error("Erro ao fechar o zipOutputStream."+e.getMessage());
+
+                zipOutputStream = new ZipOutputStream(new FileOutputStream(zipFileName + nameZip + ".zip"));
+                ZipUtil.zipFolder(new File(path), zipOutputStream, "");
+            } catch (Exception e) {
+                LOGGER.error("Erro ao tentar zipar os arquivos." + e.getMessage());
+            } finally {
+                try {
+                    zipOutputStream.close();
+                } catch (IOException e) {
+                    LOGGER.error("Erro ao fechar o zipOutputStream." + e.getMessage());
+                }
             }
+
+            try {
+                FileUtils.deleteDirectory(new File(path));
+            } catch (IOException e) {
+                LOGGER.error("Erro ao remover a pasta dos arquivos gerados para serem zipados." + e.getMessage());
+            }
+
+            return zipFileName;
         }
 
-        try {
-            FileUtils.deleteDirectory(new File(path));
-        } catch (IOException e) {
-            LOGGER.error("Erro ao remover a pasta dos arquivos gerados para serem zipados."+e.getMessage());
-        }
-
-        return zipFileName;
+        throw new RuntimeException("Nao existe arquivo");
     }
 
     private String getDirectoryPath() {
